+++ src/Notepads/Controls/TextEditor/TextEditor.xaml.cs
@@ -23,7 +23,7 @@ namespace Notepads.Controls.TextEditor
 
     public enum TextEditorMode
     {
-        Editing,
+        Editing = 0,
         DiffPreview
     }
 
@@ -32,35 +32,6 @@ namespace Notepads.Controls.TextEditor
         Untouched,
         Modified,
         RenamedMovedOrDeleted
-    }
-
-    public class TextEditorStateMetaData
-    {
-        public string LastSavedEncoding { get; set; }
-
-        public string LastSavedLineEnding { get; set; }
-
-        public long DateModifiedFileTime { get; set; }
-
-        public string RequestedLineEnding { get; set; }
-
-        public string RequestedEncoding { get; set; }
-
-        public bool HasEditingFile { get; set; }
-
-        public bool IsModified { get; set; }
-
-        public int SelectionStartPosition { get; set; }
-
-        public int SelectionEndPosition { get; set; }
-
-        public bool WrapWord { get; set; }
-
-        public double FontSize { get; set; }
-
-        public double ScrollViewerHorizontalOffset { get; set; }
-
-        public double ScrollViewerVerticalOffset { get; set; }
     }
 
     public sealed partial class TextEditor : UserControl
@@ -230,6 +201,8 @@ namespace Notepads.Controls.TextEditor
                 ScrollViewerHorizontalOffset = horizontalOffset,
                 ScrollViewerVerticalOffset = verticalOffset,
                 FontSize = TextEditorCore.FontSize,
+                IsContentPreviewPanelOpened = (SplitPanel != null && SplitPanel.Visibility == Visibility.Visible),
+                IsInDiffPreviewMode = (EditorMode == TextEditorMode.DiffPreview)
             };
 
             if (RequestedEncoding != null)
